requirements.txt
# Transit Route Analysis Project Dependencies
# Python package requirements for transit route analysis codebase

# Data manipulation and analysis
pandas>=2.0.0
geopandas>=0.14.0

# Geographic and geometric operations
shapely>=2.0.0

# OpenStreetMap network analysis
osmnx>=1.6.0

# Visualization
matplotlib>=3.7.0

# HTTP requests for Google Maps API
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0
